Dodata prikaz artikala u korpi u okviru menija. Checkout dorada (checkpoint)

+++ E-library/Services/ShoppingCart/ShoppingCart.API/Controllers/CartController.cs
@@ -39,13 +39,26 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(Cart), StatusCodes.Status200OK)]
         public async Task<ActionResult<Cart>> GetCart(string username)
         {
-            //if (User.FindFirst(ClaimTypes.Name).Value != username)
-            //{
-            //    return Forbid();
-            //}
+            if (User.FindFirst(ClaimTypes.Name).Value != username)
+            {
+                return Forbid();
+            }
 
             var cart = await _repository.GetCart(username);
             return Ok(cart ?? new Cart(username));
+        }
+        [Route("[action]/{username}")]
+        [HttpGet]
+        [ProducesResponseType(typeof(decimal), StatusCodes.Status200OK)]
+        public async Task<ActionResult<Cart>> GetCartTotalItems(string username)
+        {
+            if (User.FindFirst(ClaimTypes.Name).Value != username)
+            {
+                return Forbid();
+            }
+
+            var cart = await _repository.GetCart(username);
+            return Ok(cart.TotalItems);
         }
 
         [Route("[action]")]
@@ -53,10 +66,10 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(Cart), StatusCodes.Status200OK)]
         public async Task<ActionResult<Cart>> UpdateCart([FromBody] Cart cart)
         {
-            //if (User.FindFirst(ClaimTypes.Name).Value != cart.Username)
-            //{
-            //    return Forbid();
-            //}
+            if (User.FindFirst(ClaimTypes.Name).Value != cart.Username)
+            {
+                return Forbid();
+            }
 
             return Ok(await _repository.UpdateCart(cart));
         }
@@ -66,10 +79,10 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(void), StatusCodes.Status200OK)]
         public async Task<IActionResult> DeleteCart(string username)
         {
-            //if (User.FindFirst(ClaimTypes.Name).Value != username)
-           // {
-           //     return Forbid();
-            //}
+            if (User.FindFirst(ClaimTypes.Name).Value != username)
+            {
+                return Forbid();
+            }
 
             await _repository.DeleteCart(username);
             return Ok();
@@ -81,41 +94,25 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(void), StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> Checkout(string username)
         {
-            //if (User.FindFirst(ClaimTypes.Name).Value != cartCheckout.Username)
-            //{
-            //    return Forbid();
-            //}
+            if (User.FindFirst(ClaimTypes.Name).Value != username)
+            {
+                return Forbid();
+            }
 
-            CartCheckout cartCheckout = new CartCheckout();
-            List<Entities.CartItem> checkoutItems = new List<Entities.CartItem>();
             var cart = await _repository.GetCart(username);
             if (cart == null)
             {
                 return BadRequest("There is no active shopping cart for this user");
             }
 
-            cartCheckout.Username = username;
-           
-            cartCheckout.MemberId = null;
-            foreach (var item in cart.Items) {
-                Entities.CartItem checkoutItem = new Entities.CartItem();
-                checkoutItem.BookAuthor = item.BookAuthor;
-                checkoutItem.BookGenre = item.BookGenre;
-                checkoutItem.BookId = item.BookId;
-                checkoutItem.BookTitle = item.BookTitle;
-                checkoutItem.CoverImageFile = item.CoverImageFile;
-                checkoutItem.IsPremium = item.IsPremium;
-                checkoutItem.Language = item.Language;
-                checkoutItems.Add(checkoutItem);
-            }
-            cartCheckout.OrderItems = checkoutItems;
-
-            var eventMessage = _mapper.Map<CartCheckoutEvent>(cartCheckout);
+            var eventMessage = _mapper.Map<CartCheckoutEvent>(cart);
             await _publishEndpoint.Publish(eventMessage);
 
-            await _repository.DeleteCart(cartCheckout.Username);
+            cart.Items.Clear();
+            await _repository.UpdateCart(cart);
 
             return Accepted();
+
         }
 
         [Route("[action]/{username}/{bookId}")]
@@ -123,10 +120,10 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(Cart), StatusCodes.Status200OK)]
         public async Task<ActionResult<Cart>> AddBookToCart(string username, string bookId)
         {
-            //if (User.FindFirst(ClaimTypes.Name).Value != username)
-            //{
-            //    return Forbid();
-            //}
+            if (User.FindFirst(ClaimTypes.Name).Value != username)
+            {
+                return Forbid();
+            }
 
             var result =  await _service.AddBookToCart(username, bookId);
             if (result == null)
@@ -150,10 +147,10 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(Cart), StatusCodes.Status200OK)]
         public async Task<ActionResult<Cart>> RemoveBookFromCart(string username, string bookId)
         {
-            //if (User.FindFirst(ClaimTypes.Name).Value != username)
-            //{
-            //    return Forbid();
-            //}
+            if (User.FindFirst(ClaimTypes.Name).Value != username)
+            {
+                return Forbid();
+            }
 
             return Ok(await _service.RemoveBookFromCart(username, bookId));
         }
